+++ changelog.md
@@ -1,2 +1,1 @@
-Add sort order for category annotations.
-Make config screen a pause screen. 
+use different comment parsing system