+++ changelog.md
@@ -1,1 +1,1 @@
-Fix crash on neoforge
+use different comment parsing system