+++ changelog.md
@@ -1,3 +1,1 @@
-- Update to 1.20.5
-- Add config versioning and config patching.
-- Fix file name not showing up correctly in the UI.
+use different comment parsing system