+++ changelog.md
@@ -1,1 +1,1 @@
-- Fixed incorrect java version in mixin config
+use different comment parsing system